Исправить мелкие баги и несоответствия после структурного рефакторинга клиентской части RControl Web

+++ Core/Dtos/RControl/Lookups/BillingPeriodDto.cs
@@ -2,4 +2,4 @@
 
 public sealed record BillingPeriodDto(
     int BillingYear,
-    byte BillingMonth);
+    int BillingMonth);
